Move conf back to /var/acme to keep files during upgrades

+++ setup.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env sh
 # Setup
 HOME=/var/acme
-CONF=/conf/acme
+CONF=/var/acme
 mkdir -p $HOME/http
 mkdir -p $CONF
 cd $HOME
@@ -9,7 +9,10 @@ curl https://raw.githubusercontent.com/acmesh-official/acme.sh/master/acme.sh -o
 chmod +x acme.sh
 curl https://raw.githubusercontent.com/helsinki-systems/sophos-xgs-letsencrypt/main/renew.sh -o renew.sh
 chmod +x renew.sh
-curl https://raw.githubusercontent.com/helsinki-systems/sophos-xgs-letsencrypt/main/config.csv -o $CONF/config.csv
+if [ -e /var/acme/config.csv ]
+then
+    curl https://raw.githubusercontent.com/helsinki-systems/sophos-xgs-letsencrypt/main/config.csv -o $CONF/config.csv
+fi
 echo "Please insert letsencrypt account email"
 read email
 chmod +x $HOME/acme.sh
